requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
lightgbm
catboost
matplotlib
seaborn
plotly
google-generativeai
python-dotenv